+++ src/Map/src/Polygon.php
@@ -22,7 +22,9 @@ final class Polygon implements Element
 {
     /**
      * @param array<Point>|array<array<Point>> $points a list of point representing the polygon, or a list of paths (each path is an array of points) representing a polygon with holes
-     * @param array<string, mixed>             $extra  Extra data, can be used by the developer to store additional information and use them later JavaScript side
+     * @param array<string, mixed>             $extra  Extra data forwarded to the JavaScript side. It can be used in your custom
+     *                                                 Stimulus controller to benefit from greater flexibility and customization.
+     *                                                 These data must be serializable to JSON. These data are not used by UX Map.
      */
     public function __construct(
         private readonly array $points,