+++ src/Map/src/Polyline.php
@@ -21,7 +21,9 @@ use Symfony\UX\Map\Exception\InvalidArgumentException;
 final class Polyline implements Element
 {
     /**
-     * @param array<string, mixed> $extra Extra data, can be used by the developer to store additional information and use them later JavaScript side
+     * @param array<string, mixed> $extra Extra data forwarded to the JavaScript side. It can be used in your custom
+     *                                    Stimulus controller to benefit from greater flexibility and customization.
+     *                                    These data must be serializable to JSON. These data are not used by UX Map.
      */
     public function __construct(
         private readonly array $points,